requirements.txt
streamlit
transformers
torch
pillow
deepface
tf-keras
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
numpy
reportlab
